requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.1
Automat==22.10.0
billiard==3.6.4.0
celery==5.2.7
cffi==1.15.1
channels==4.0.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
constantly==15.1.0
cryptography==39.0.0
daphne==4.0.0
Django==4.1.3
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-redis==5.2.0
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
eventlet==0.33.3
greenlet==2.0.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
kombu==5.2.4
msgpack==1.0.4
Pillow==9.3.0
prompt-toolkit==3.0.36
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.0.0
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2022.6
redis==4.4.2
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==23.1.1
typing_extensions==4.4.0
tzdata==2022.6
vine==5.0.0
wcwidth==0.2.6
zope.interface==5.5.2
